LightGBM.Model

Models

data Model

A model to use to make predictions

trainNewModel

Arguments

:: [Param]
Training parameters

-> DataSet
Training data

-> [DataSet]
Testing data

-> IO (Either ErrLog Model)

Train a new model and persist it to a file.

readModelFile :: FilePath -> IO Model

Persisted models can be loaded up and used for prediction.

writeModelFile :: FilePath -> Model -> IO ()

Models can be written out to a file

Prediction

predict

Arguments

:: Model
A model to do prediction with

-> [Param]
Prediction parameters

-> DataSet
The new input data for prediction

-> IO (Either ErrLog DataSet)
The prediction output DataSet

Predict the results of new inputs and persist the results to an output file.