LightGBM.Parameters

Description

Parameter types for LightGBM.

Parameter details are documented in the LightGBM documentation.

Note that some of the parameters listed in the documentation are not exposed here since they're set implicitly through other parts of the API. For instance, the task param is set in the LightGBM.Model API, and the header param is set in the LightGBM.DataSet API.

Parameters

data Param

Parameters control the behavior of lightGBM.

Constructors

Objective Application
Regression, binary classification, etc.

BoostingType Booster
Booster to apply - by default is GBDT

TrainingData FilePath
Path to training data

ValidationData [FilePath]
Paths to validation data files (supports multi-validation)

PredictionData FilePath
Path to data to use for a prediction

Iterations Natural
Number of boosting iterations - default is 100

LearningRate PositiveDouble
Scale how quickly parameters change in training

NumLeaves PositiveInt
Maximum number of leaves in one tree

Parallelism ParallelismStyle
Called tree_learner in the LightGBM docs

NumThreads Natural
Number of threads for LightGBM to use

Device Device
GPU or CPU

RandomSeed Int
A random seed used to generate other random seeds

MaxDepth Natural
Limit the depth of the tree model

MinDataInLeaf Natural
Minimum data count in a leaf

MinSumHessianInLeaf NonNegativeDouble
Minimal sum of the Hessian in one leaf

BaggingFraction LeftOpenProperFraction

BaggingFreq PositiveInt

BaggingFractionSeed Int

FeatureFraction LeftOpenProperFraction

FeatureFractionSeed Int

EarlyStoppingRounds PositiveInt
Stop training if a validation metric doesn't improve in the last n rounds

Regularization_L1 NonNegativeDouble

Regularization_L2 NonNegativeDouble

MaxDeltaStep PositiveDouble

MinSplitGain NonNegativeDouble

MinDataPerGroup PositiveInt
Minimum number of data points per categorial group

MaxCatThreshold PositiveInt

CatSmooth NonNegativeDouble

CatL2 NonNegativeDouble
L2 regularization in categorical split

MaxCatToOneHot PositiveInt

TopK PositiveInt
VotingPar only

MonotoneConstraint [Direction]
Length of directions = number of features

MaxBin IntGreaterThanOne

MinDataInBin PositiveInt

DataRandomSeed Int

OutputModel FilePath
Where to persist the model after training

InputModel FilePath
Filepath to a persisted model to use for prediction or additional training

OutputResult FilePath
Where to persist the output of a prediction task

PrePartition Bool

IsSparse Bool

TwoRoundLoading Bool

SaveBinary Bool

Verbosity VerbosityLevel

LabelColumn ColumnSelector
Which column has the labels

WeightColumn ColumnSelector
Which column has the weights

QueryColumn ColumnSelector

IgnoreColumns [ColumnSelector]
Select columns to ignore in training

CategoricalFeatures [ColumnSelector]
Select columns to use as features

PredictRawScore Bool
Prediction Only; true = raw scores only, false = transformed scores

PredictLeafIndex Bool
Prediction Only

PredictContrib Bool
Prediction Only

BinConstructSampleCount PositiveInt

NumIterationsPredict Natural
Prediction Only; how many trained predictions

PredEarlyStop Bool

PredEarlyStopFreq Natural

PredEarlyStopMargin Double

UseMissing Bool

ZeroAsMissing Bool

InitScoreFile FilePath

ValidInitScoreFile [FilePath]

ForcedSplits FilePath

Sigmoid PositiveDouble
Used in Binary classification and LambdaRank

Alpha OpenProperFraction
Used in Huber loss and Quantile regression

BoostFromAverage Bool
Used only in RegressionL2 task

RegSqrt Bool
Only used in RegressionL2

Metric [Metric]
Loss Metric

MetricFreq PositiveInt

TrainingMetric Bool

data Application

LightGBM can be used for a variety of applications

Constructors

Regression RegressionApp

BinaryClassification [BinaryClassParam]

MultiClass MultiClassStyle NumClasses

CrossEntropy XEApp

LambdaRank [LambdaRankParam]
A ranking algorithm

data BinaryClassParam

Constructors

IsUnbalance Bool
Set to true if training data are unbalanced

ScalePosWeight PositiveDouble

data Booster

Different types of Boosting approaches

Constructors

GBDT
Gradient Boosting Decision Tree

RandomForest

DART [DARTParam]
Dropouts meet Multiple Additive Regression Trees

GOSS [GOSSParam]
Gradient-based One-Sided Sampling

data DARTParam

Parameters exclusively for the DART booster

Constructors

DropRate ProperFraction
Dropout rate

SkipDrop ProperFraction
Probablility of skipping a drop

MaxDrop PositiveInt
Max number of dropped trees on one iteration

UniformDrop Bool

XGBoostDARTMode Bool

DropSeed Int

data Device

Constructors

CPU

GPU [GPUParam]

data Direction

Constructors

Increasing

Decreasing

NoConstraint

data FairRegressionParam

Constructors

FairC PositiveDouble

data GOSSParam

Constructors

TopRate ProperFraction

OtherRate ProperFraction

data GPUParam

Constructors

GpuPlatformId Natural

GpuDeviceId Natural

GpuUseDP Bool

data LambdaRankParam

Constructors

MaxPosition PositiveInt

LabelGain [Double]
Used in LambdaRank

type LocalListenPort = Natural

type MachineListFile = FilePath

data Metric

Constructors

MeanAbsoluteError

MeanSquareError

L2_root

QuantileRegression

MAPELoss

HuberLoss

FairLoss

PoissonNegLogLikelihood

GammaNegLogLikelihood

GammaDeviance

TweedieNegLogLiklihood

NDCG (Maybe NDCGEvalPositions)

MAP

AUC

BinaryLogloss

BinaryError

MultiLogloss

MultiError

Xentropy

XentLambda

KullbackLeibler

type Minutes = Natural

data MultiClassStyle

Multi-classification styles

Constructors

MultiClassSimple

MultiClassOneVsAll

type NDCGEvalPositions = [Natural]

type NumClasses = Natural

type NumMachines = PositiveInt

data ParallelismParams

Constructors

SocketVer

Fields

nMachines :: NumMachines

machineList :: MachineListFile

port :: LocalListenPort

timeOut :: Minutes

MPIVer

Fields

nMachines :: NumMachines

data ParallelismStyle

Constructors

Serial

FeaturePar ParallelismParams

DataPar ParallelismParams

VotingPar ParallelismParams

data PoissonRegressionParam

Parameters for Poisson regression

Constructors

PoissonMaxDeltaStep PositiveDouble

data RegressionApp

Different types of regression metrics

Constructors

L1
Absolute error metric

L2
RMS errror metric

Huber

Fair [FairRegressionParam]

Poisson [PoissonRegressionParam]

Quantile

MAPE

Gamma

Tweedie [TweedieRegressionParam]

data TweedieRegressionParam

Constructors

TweedieVariancePower OneToTwoLeftSemiClosed
Control Tweedie variance in range [1, 2) - 1 is like Poisson, 2 is like Gamma

data VerbosityLevel

Constructors

Fatal

Warn

Info

data XEApp

Constructors

XEntropy

XEntropyLambda

Utilities

data ColumnSelector

Some parameters are based on column selection either by index or by name. A ColumnSelector encapsulates this flexibility.

Constructors

Index Natural

ColName String

colSelArgument :: ColumnSelector -> String